@@ -7,14 +7,19 @@ using System.Threading.Tasks;
 using ITSolution.Framework.Core.BaseClasses;
 using ITSolution.Framework.Server.Core.BaseClasses.Repository;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Identity;
+//using Microsoft.AspNetCore.Identity.UI;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.HttpsPolicy;
+using IHostingEnvironment = Microsoft.Extensions.Hosting.IHostingEnvironment;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using ITSolution.Framework.Core.Server.BaseClasses.Repository.Identity;
+using ITSolution.Framework.BaseClasses;
 
 namespace ITSolution.Framework.Core.Server.BaseClasses
 {
@@ -30,18 +35,25 @@ namespace ITSolution.Framework.Core.Server.BaseClasses
         {
             Configuration = configuration;
             //assinando o evento para definir quem vai resolver os assemblies
-            AssemblyLoadContext.Default.Resolving += Default_Resolving;
+            //AssemblyLoadContext.Default.Resolving += Default_Resolving;
         }
 
         private Assembly Default_Resolving(AssemblyLoadContext arg1, AssemblyName arg2)
         {
-            return ITSAssemblyResolve.ITSLoader.LoadFromAssemblyName(arg2);
+            if (arg2.Name.Contains("ITSolution"))
+                return ITSAssemblyResolve.ITSLoader.LoadFromAssemblyName(arg2);
+            else
+            {
+                AssemblyLoadContext ctx = new AssemblyLoadContext("StandardNET", false);
+                return ctx.LoadFromAssemblyName(arg2);
+            }
         }
 
         protected virtual IConfiguration Configuration { get; }
 
         public IServiceProvider ConfigureServices(IServiceCollection services)
         {
+            services.ConfigureIdentity();
             services.Configure<CookiePolicyOptions>(options =>
             {
                 // This lambda determines whether user consent for non-essential cookies is needed for a given request.
@@ -49,9 +61,10 @@ namespace ITSolution.Framework.Core.Server.BaseClasses
                 options.MinimumSameSitePolicy = SameSiteMode.None;
             });
 
-            IMvcBuilder mvcBuilder = services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
             try
             {
+                services.AddDatabaseDeveloperPageExceptionFilter();
+                IMvcBuilder mvcBuilder = services.AddMvc();
                 //starting application parts
                 foreach (var file in ITSAssemblyResolve.ITSLoader.GetServerAssemblies())
                 {
@@ -63,35 +76,50 @@ namespace ITSolution.Framework.Core.Server.BaseClasses
                 {
                     services.Replace(item);
                 }
-                
+
+                services.AddRazorPages();
                 return services.BuildServiceProvider();
             }
             catch (Exception ex)
             {
                 Utils.ShowExceptionStack(ex);
-                throw ex;
+                throw;
             }
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app)
         {
-            var env = app.ApplicationServices.GetService<IHostingEnvironment>();
+            var env = app.ApplicationServices.GetService<IHostEnvironment>();
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseMigrationsEndPoint();
             }
             else
             {
                 app.UseExceptionHandler("/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
 
             app.UseHttpsRedirection();
             app.UseStaticFiles();
-            app.UseCookiePolicy();
 
-            app.UseMvc();
+            app.UseRouting();
+
+
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapRazorPages();
+                endpoints.MapControllers();
+
+            });
+
         }
     }
 }
