@@ -120,10 +120,10 @@ namespace ITSolution.Framework.Core.BaseClasses
                 DirectoryInfo dir = new DirectoryInfo(tarjet);
                 return dir.GetFiles();
             }
-            catch (Exception ex)
+            catch (Exception )
             {
 
-                throw ex;
+                throw ;
             }
         }
 
@@ -137,10 +137,10 @@ namespace ITSolution.Framework.Core.BaseClasses
                 return dir.GetDirectories();
 
             }
-            catch (Exception ex)
+            catch (Exception )
             {
 
-                throw ex;
+                throw ;
             }
         }
 
@@ -196,7 +196,7 @@ namespace ITSolution.Framework.Core.BaseClasses
             {
                 string msg = string.Format("Falha ao abrir o arquivo {0}\n{1}", pathFile, ex.Message);
                 Console.WriteLine(msg);
-                throw ex;
+                throw;
             }
             return result;
         }
@@ -291,7 +291,7 @@ namespace ITSolution.Framework.Core.BaseClasses
             {
                 string msg = string.Format("Falha ao abrir o arquivo {0}\n{1}", pathFile, ex.Message);
                 Trace.WriteLine(msg);
-                throw ex;
+                throw ;
             }
             finally
             {
@@ -553,9 +553,9 @@ namespace ITSolution.Framework.Core.BaseClasses
                 else
                     return false;
             }
-            catch (System.IO.FileNotFoundException ex)
+            catch (System.IO.FileNotFoundException )
             {
-                throw ex;
+                throw ;
             }
         }
 
@@ -836,9 +836,9 @@ namespace ITSolution.Framework.Core.BaseClasses
                 return File.ReadAllBytes(path);
 
             }
-            catch (Exception ex)
+            catch (Exception )
             {
-                throw ex;
+                throw ;
             }
         }
 
@@ -1075,7 +1075,7 @@ namespace ITSolution.Framework.Core.BaseClasses
             catch (Exception ex)
             {
                 Utils.ShowExceptionStack(ex);
-                throw ex;
+                throw;
             }
             return fileList;
         }
@@ -1157,7 +1157,7 @@ namespace ITSolution.Framework.Core.BaseClasses
             catch (Exception ex)
             {
                 Utils.ShowExceptionStack(ex);
-                throw ex;
+                throw ;
             }
 
         }
@@ -1180,7 +1180,7 @@ namespace ITSolution.Framework.Core.BaseClasses
             catch (Exception ex)
             {
                 Utils.ShowExceptionStack(ex);
-                throw ex;
+                throw;
             }
 
         }
@@ -1230,7 +1230,7 @@ namespace ITSolution.Framework.Core.BaseClasses
             {
                 Console.WriteLine("Falha ao mover " + alvo + " para " + destino);
                 Utils.ShowExceptionStack(ex);
-                throw ex;
+                throw ;
             }
         }
 
