@@ -1,5 +1,6 @@
 ﻿using ITSolution.Framework.BaseClasses;
 using Oracle.ManagedDataAccess.Client;
+//using Oracle.ManagedDataAccess.Client;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -14,7 +15,7 @@ namespace ITSolution.Framework.Core.Server.BaseClasses.Repository
         public static void ConfigureDataSources()
         {
             OracleConfiguration.OracleDataSources.Add(EnvironmentInformation.DefaultConnectionName, EnvironmentInformation.TnsAtpSource);
-            //OracleConfiguration.WalletLocation = EnvironmentInformation.WalletLocation;
+            OracleConfiguration.WalletLocation = EnvironmentInformation.WalletLocation;
         }
     }
 }
